erlang: build OTP from git repository
This aims to make building newer versions easier.

Makefile changes:
 * OTP_VERSION is either a tag or commit id in the otp repository.
 * OTP_DIRECTORY is introduced for readability.

Patches are moved into a patches directory. The patches are added to the
git repository using 'git am', and extracted using 'git format-patch'.
This allows for commenting in the patch. At the moment the patches usage
is unknown. There is also a guarantee that the patches will be added
smoothly, but also easier handling when mering the patches.

A problem when using --without-wx flag and git. Somehoe this crashed the
build for a number of different libraries inside OTP. This is solved by
skipping this libraries too.